Split Ghostty surface view helpers
