+++ supacode/Infrastructure/Ghostty/GhosttySurfaceView.swift
@@ -5,7 +5,7 @@ import GhosttyKit
 import QuartzCore
 import SwiftUI
 
-private let surfaceLogger = SupaLogger("Surface")
+let surfaceLogger = SupaLogger("Surface")
 
 enum GhosttyEventText {
   static func characters(for event: NSEvent) -> String? {
@@ -80,7 +80,7 @@ final class GhosttySurfaceView: NSView, Identifiable {
     }
   }
 
-  private final class CachedValue<T> {
+  final class CachedValue<T> {
     private var value: T?
     private let fetch: () -> T
     private let duration: Duration
@@ -114,7 +114,7 @@ final class GhosttySurfaceView: NSView, Identifiable {
     }
   }
 
-  private let runtime: GhosttyRuntime
+  let runtime: GhosttyRuntime
   let id = UUID()
   private var debugID: String {
     String(id.uuidString.prefix(8))
@@ -138,22 +138,22 @@ final class GhosttySurfaceView: NSView, Identifiable {
   private let skipsSurfaceCreationForTesting: Bool
   private var trackingArea: NSTrackingArea?
   private var lastBackingSize: CGSize = .zero
-  private var lastPerformKeyEvent: TimeInterval?
+  var lastPerformKeyEvent: TimeInterval?
   private var currentCursor: NSCursor = .iBeam
-  private var focused = false
+  var focused = false
   private var detachedFocusClearTask: Task<Void, Never>?
-  private var markedText = NSMutableAttributedString()
-  private var keyboardLayoutChangeKeyUpSuppression: KeyboardLayoutChangeKeyUpSuppression?
-  private var keyTextAccumulator: [String]?
-  private var cellSize: CGSize = .zero
+  var markedText = NSMutableAttributedString()
+  var keyboardLayoutChangeKeyUpSuppression: KeyboardLayoutChangeKeyUpSuppression?
+  var keyTextAccumulator: [String]?
+  var cellSize: CGSize = .zero
   private var lastScrollbar: ScrollbarState?
   private var occlusionState = OcclusionState()
   private var lastSurfaceFocus: Bool?
   private var eventMonitor: Any?
   private var notificationObservers: [NSObjectProtocol] = []
-  private var prevPressureStage: Int = 0
+  var prevPressureStage: Int = 0
   private var isBackgroundOpaqueOverride = false
-  private lazy var cachedScreenContents = CachedValue<String>(duration: .milliseconds(500)) {
+  lazy var cachedScreenContents = CachedValue<String>(duration: .milliseconds(500)) {
     [weak self] in
     self?.readScreenContents() ?? ""
   }
@@ -187,7 +187,7 @@ final class GhosttySurfaceView: NSView, Identifiable {
   var onOcclusionAppliedForTesting: ((Bool) -> Void)?
   var attachmentStateForTesting: (() -> (hasSuperview: Bool, hasWindow: Bool))?
 
-  private var accessibilityPaneIndexHelp: String?
+  var accessibilityPaneIndexHelp: String?
 
   private static let mouseCursorMap: [ghostty_action_mouse_shape_e: NSCursor] = [
     GHOSTTY_MOUSE_SHAPE_DEFAULT: .arrow,
@@ -214,7 +214,7 @@ final class GhosttySurfaceView: NSView, Identifiable {
     GHOSTTY_MOUSE_SHAPE_S_RESIZE,
     GHOSTTY_MOUSE_SHAPE_NS_RESIZE,
   ]
-  private static let dropTypes: Set<NSPasteboard.PasteboardType> = [
+  static let dropTypes: Set<NSPasteboard.PasteboardType> = [
     .string,
     .fileURL,
     .URL,
@@ -245,7 +245,7 @@ final class GhosttySurfaceView: NSView, Identifiable {
     return String(bytes: buffer, encoding: .utf8) ?? ""
   }
 
-  private static func string(from text: ghostty_text_s) -> String {
+  static func string(from text: ghostty_text_s) -> String {
     stringFromGhosttyText(pointer: text.text, length: text.text_len)
   }
 
@@ -563,424 +563,6 @@ final class GhosttySurfaceView: NSView, Identifiable {
     }
   }
 
-  func setAccessibilityPaneIndex(index: Int, total: Int) {
-    guard total > 0, index > 0, index <= total else {
-      accessibilityPaneIndexHelp = nil
-      return
-    }
-    accessibilityPaneIndexHelp = "Pane \(index) of \(total)"
-  }
-
-  override func isAccessibilityElement() -> Bool {
-    // Avoid interacting with panes after teardown.
-    surface != nil
-  }
-
-  override func accessibilityRole() -> NSAccessibility.Role? {
-    // Match Ghostty.app so speech/input tools can treat the surface as editable text.
-    .textArea
-  }
-
-  override func accessibilityLabel() -> String? {
-    let title = bridge.state.title?.trimmingCharacters(in: .whitespacesAndNewlines) ?? ""
-    if !title.isEmpty {
-      return title
-    }
-    let pwd = bridge.state.pwd?.trimmingCharacters(in: .whitespacesAndNewlines) ?? ""
-    if !pwd.isEmpty {
-      return pwd
-    }
-    return "Terminal pane"
-  }
-
-  override func accessibilityValue() -> Any? {
-    cachedScreenContents.get()
-  }
-
-  override func accessibilityHelp() -> String? {
-    accessibilityPaneIndexHelp
-  }
-
-  override func accessibilitySelectedTextRange() -> NSRange {
-    selectedRange()
-  }
-
-  override func accessibilitySelectedText() -> String? {
-    guard let surface else { return nil }
-    var text = ghostty_text_s()
-    guard ghostty_surface_read_selection(surface, &text) else { return nil }
-    defer { ghostty_surface_free_text(surface, &text) }
-    let value = Self.string(from: text)
-    return value.isEmpty ? nil : value
-  }
-
-  override func accessibilityNumberOfCharacters() -> Int {
-    cachedScreenContents.get().count
-  }
-
-  override func accessibilityVisibleCharacterRange() -> NSRange {
-    let content = cachedScreenContents.get()
-    return NSRange(location: 0, length: content.count)
-  }
-
-  override func accessibilityLine(for index: Int) -> Int {
-    Self.accessibilityLine(for: index, in: cachedScreenContents.get())
-  }
-
-  override func accessibilityString(for range: NSRange) -> String? {
-    Self.accessibilityString(for: range, in: cachedScreenContents.get())
-  }
-
-  override func accessibilityAttributedString(for range: NSRange) -> NSAttributedString? {
-    guard let surface else { return nil }
-    guard let plainString = accessibilityString(for: range) else { return nil }
-
-    var attributes: [NSAttributedString.Key: Any] = [:]
-    if let fontRaw = ghostty_surface_quicklook_font(surface) {
-      let font = Unmanaged<CTFont>.fromOpaque(fontRaw)
-      attributes[.font] = font.takeUnretainedValue()
-      font.release()
-    }
-
-    return NSAttributedString(string: plainString, attributes: attributes)
-  }
-
-  override func becomeFirstResponder() -> Bool {
-    let result = super.becomeFirstResponder()
-    if result {
-      focusDidChange(true)
-      postAccessibilityFocusChanged()
-    }
-    return result
-  }
-
-  override func resignFirstResponder() -> Bool {
-    let result = super.resignFirstResponder()
-    if result {
-      focusDidChange(false)
-    }
-    return result
-  }
-
-  private func postAccessibilityFocusChanged() {
-    guard surface != nil else { return }
-    // Post on the window so assistive tech can query the focused element from it.
-    if let window {
-      NSAccessibility.post(element: window, notification: .focusedUIElementChanged)
-    } else {
-      NSAccessibility.post(element: self, notification: .focusedUIElementChanged)
-    }
-  }
-
-  private func readText(
-    topLeftTag: ghostty_point_tag_e,
-    bottomRightTag: ghostty_point_tag_e
-  ) -> String? {
-    guard let surface else { return nil }
-    var text = ghostty_text_s()
-    let selection = ghostty_selection_s(
-      top_left: ghostty_point_s(
-        tag: topLeftTag,
-        coord: GHOSTTY_POINT_COORD_TOP_LEFT,
-        x: 0,
-        y: 0
-      ),
-      bottom_right: ghostty_point_s(
-        tag: bottomRightTag,
-        coord: GHOSTTY_POINT_COORD_BOTTOM_RIGHT,
-        x: 0,
-        y: 0
-      ),
-      rectangle: false
-    )
-    guard ghostty_surface_read_text(surface, selection, &text) else { return nil }
-    defer { ghostty_surface_free_text(surface, &text) }
-    return Self.string(from: text)
-  }
-
-  private func readScreenContents() -> String {
-    readText(
-      topLeftTag: GHOSTTY_POINT_SCREEN,
-      bottomRightTag: GHOSTTY_POINT_SCREEN
-    ) ?? ""
-  }
-
-  func readViewportContentsForCLI() -> String? {
-    readText(
-      topLeftTag: GHOSTTY_POINT_VIEWPORT,
-      bottomRightTag: GHOSTTY_POINT_VIEWPORT
-    )
-  }
-
-  func readActiveContentsForCLI() -> String? {
-    readText(
-      topLeftTag: GHOSTTY_POINT_ACTIVE,
-      bottomRightTag: GHOSTTY_POINT_ACTIVE
-    )
-  }
-
-  func readScreenContentsForCLI() -> String? {
-    readText(
-      topLeftTag: GHOSTTY_POINT_SCREEN,
-      bottomRightTag: GHOSTTY_POINT_SCREEN
-    )
-  }
-
-  override func keyDown(with event: NSEvent) {
-    guard let surface else {
-      interpretKeyEvents([event])
-      return
-    }
-    bridge.state.bellCount = 0
-    onKeyInput?()
-    if let mirroredKey = MirroredTerminalKey(event: event) {
-      onMirroredKey?(mirroredKey)
-    }
-    let (translationEvent, translationMods) = translationState(event, surface: surface)
-    let action = event.isARepeat ? GHOSTTY_ACTION_REPEAT : GHOSTTY_ACTION_PRESS
-    keyTextAccumulator = []
-    defer { keyTextAccumulator = nil }
-    let markedTextBefore = markedText.length > 0
-    let keyboardIdBefore = markedTextBefore ? nil : keyboardLayoutId()
-    lastPerformKeyEvent = nil
-    interpretKeyEvents([translationEvent])
-    if !markedTextBefore, keyboardIdBefore != keyboardLayoutId() {
-      keyboardLayoutChangeKeyUpSuppression = KeyboardLayoutChangeKeyUpSuppression(
-        keyCode: event.keyCode,
-        timestamp: event.timestamp
-      )
-      return
-    }
-    syncPreedit(clearIfNeeded: markedTextBefore)
-    if let list = keyTextAccumulator, !list.isEmpty {
-      for text in list {
-        _ = sendKey(
-          action: action,
-          event: event,
-          translationEvent: translationEvent,
-          translationMods: translationMods,
-          text: text,
-          composing: false
-        )
-      }
-    } else {
-      _ = sendKey(
-        action: action,
-        event: event,
-        translationEvent: translationEvent,
-        translationMods: translationMods,
-        text: ghosttyCharacters(translationEvent),
-        composing: markedText.length > 0 || markedTextBefore
-      )
-    }
-  }
-
-  override func keyUp(with event: NSEvent) {
-    if suppressKeyboardLayoutChangeKeyUp(event) { return }
-    sendKey(action: GHOSTTY_ACTION_RELEASE, event: event)
-  }
-
-  override func flagsChanged(with event: NSEvent) {
-    let mod: UInt32
-    switch event.keyCode {
-    case 0x39: mod = GHOSTTY_MODS_CAPS.rawValue
-    case 0x38, 0x3C: mod = GHOSTTY_MODS_SHIFT.rawValue
-    case 0x3B, 0x3E: mod = GHOSTTY_MODS_CTRL.rawValue
-    case 0x3A, 0x3D: mod = GHOSTTY_MODS_ALT.rawValue
-    case 0x37, 0x36: mod = GHOSTTY_MODS_SUPER.rawValue
-    default: return
-    }
-    if hasMarkedText() { return }
-    let mods = ghosttyMods(event.modifierFlags)
-    var action = GHOSTTY_ACTION_RELEASE
-    if (mods.rawValue & mod) != 0 {
-      let sidePressed: Bool
-      switch event.keyCode {
-      case 0x3C:
-        sidePressed = event.modifierFlags.rawValue & UInt(NX_DEVICERSHIFTKEYMASK) != 0
-      case 0x3E:
-        sidePressed = event.modifierFlags.rawValue & UInt(NX_DEVICERCTLKEYMASK) != 0
-      case 0x3D:
-        sidePressed = event.modifierFlags.rawValue & UInt(NX_DEVICERALTKEYMASK) != 0
-      case 0x36:
-        sidePressed = event.modifierFlags.rawValue & UInt(NX_DEVICERCMDKEYMASK) != 0
-      default:
-        sidePressed = true
-      }
-      if sidePressed {
-        action = GHOSTTY_ACTION_PRESS
-      }
-    }
-    sendKey(action: action, event: event)
-  }
-
-  override func mouseMoved(with event: NSEvent) {
-    sendMousePosition(event)
-    if let window, window.isKeyWindow, !focused, runtime.focusFollowsMouse() {
-      requestFocus()
-    }
-  }
-
-  override func mouseEntered(with event: NSEvent) {
-    super.mouseEntered(with: event)
-    sendMousePosition(event)
-  }
-
-  override func mouseExited(with event: NSEvent) {
-    if NSEvent.pressedMouseButtons != 0 {
-      return
-    }
-    guard let surface else { return }
-    let mods = ghosttyMods(event.modifierFlags)
-    ghostty_surface_mouse_pos(surface, -1, -1, mods)
-  }
-
-  override func mouseDown(with event: NSEvent) {
-    sendMouseButton(event, state: GHOSTTY_MOUSE_PRESS, button: GHOSTTY_MOUSE_LEFT)
-  }
-
-  override func mouseUp(with event: NSEvent) {
-    prevPressureStage = 0
-    sendMouseButton(event, state: GHOSTTY_MOUSE_RELEASE, button: GHOSTTY_MOUSE_LEFT)
-    if let surface {
-      ghostty_surface_mouse_pressure(surface, 0, 0)
-    }
-  }
-
-  override func rightMouseDown(with event: NSEvent) {
-    guard let surface else {
-      super.rightMouseDown(with: event)
-      return
-    }
-    let mods = ghosttyMods(event.modifierFlags)
-    if ghostty_surface_mouse_button(surface, GHOSTTY_MOUSE_PRESS, GHOSTTY_MOUSE_RIGHT, mods) {
-      return
-    }
-    super.rightMouseDown(with: event)
-  }
-
-  override func rightMouseUp(with event: NSEvent) {
-    guard let surface else {
-      super.rightMouseUp(with: event)
-      return
-    }
-    let mods = ghosttyMods(event.modifierFlags)
-    if ghostty_surface_mouse_button(surface, GHOSTTY_MOUSE_RELEASE, GHOSTTY_MOUSE_RIGHT, mods) {
-      return
-    }
-    super.rightMouseUp(with: event)
-  }
-
-  override func otherMouseDown(with event: NSEvent) {
-    sendMouseButton(
-      event, state: GHOSTTY_MOUSE_PRESS, button: Self.ghosttyMouseButton(from: event.buttonNumber))
-  }
-
-  override func otherMouseUp(with event: NSEvent) {
-    sendMouseButton(
-      event, state: GHOSTTY_MOUSE_RELEASE, button: Self.ghosttyMouseButton(from: event.buttonNumber)
-    )
-  }
-
-  private static func ghosttyMouseButton(from buttonNumber: Int) -> ghostty_input_mouse_button_e {
-    switch buttonNumber {
-    case 0: GHOSTTY_MOUSE_LEFT
-    case 1: GHOSTTY_MOUSE_RIGHT
-    case 2: GHOSTTY_MOUSE_MIDDLE
-    case 3: GHOSTTY_MOUSE_EIGHT
-    case 4: GHOSTTY_MOUSE_NINE
-    case 5: GHOSTTY_MOUSE_SIX
-    case 6: GHOSTTY_MOUSE_SEVEN
-    case 7: GHOSTTY_MOUSE_FOUR
-    case 8: GHOSTTY_MOUSE_FIVE
-    case 9: GHOSTTY_MOUSE_TEN
-    case 10: GHOSTTY_MOUSE_ELEVEN
-    default: GHOSTTY_MOUSE_UNKNOWN
-    }
-  }
-
-  override func mouseDragged(with event: NSEvent) {
-    sendMousePosition(event)
-  }
-
-  override func rightMouseDragged(with event: NSEvent) {
-    sendMousePosition(event)
-  }
-
-  override func otherMouseDragged(with event: NSEvent) {
-    sendMousePosition(event)
-  }
-
-  override func scrollWheel(with event: NSEvent) {
-    guard let surface else { return }
-    var scrollX = event.scrollingDeltaX
-    var scrollY = event.scrollingDeltaY
-    if event.hasPreciseScrollingDeltas {
-      scrollX *= 2
-      scrollY *= 2
-    }
-    ghostty_surface_mouse_scroll(surface, scrollX, scrollY, scrollMods(for: event))
-  }
-
-  override func pressureChange(with event: NSEvent) {
-    guard let surface else { return }
-    ghostty_surface_mouse_pressure(surface, UInt32(event.stage), Double(event.pressure))
-    guard prevPressureStage < 2 else { return }
-    prevPressureStage = event.stage
-    guard event.stage == 2 else { return }
-    guard UserDefaults.standard.bool(forKey: "com.apple.trackpad.forceClick") else { return }
-    quickLook(with: event)
-  }
-
-  override func quickLook(with event: NSEvent) {
-    guard let surface else { return super.quickLook(with: event) }
-    var text = ghostty_text_s()
-    guard ghostty_surface_quicklook_word(surface, &text) else {
-      return super.quickLook(with: event)
-    }
-    defer { ghostty_surface_free_text(surface, &text) }
-    guard text.text_len > 0 else { return super.quickLook(with: event) }
-
-    var attributes: [NSAttributedString.Key: Any] = [:]
-    if let fontRaw = ghostty_surface_quicklook_font(surface) {
-      let font = Unmanaged<CTFont>.fromOpaque(fontRaw)
-      attributes[.font] = font.takeUnretainedValue()
-      font.release()
-    }
-
-    let str = NSAttributedString(string: Self.string(from: text), attributes: attributes)
-    let point = NSPoint(x: text.tl_px_x, y: frame.size.height - text.tl_px_y)
-    showDefinition(for: str, at: point)
-  }
-
-  private func localEventHandler(_ event: NSEvent) -> NSEvent? {
-    switch event.type {
-    case .keyUp:
-      localEventKeyUp(event)
-    case .leftMouseDown:
-      localEventLeftMouseDown(event)
-    default:
-      event
-    }
-  }
-
-  private func localEventKeyUp(_ event: NSEvent) -> NSEvent? {
-    if !event.modifierFlags.contains(.command) { return event }
-    guard focused else { return event }
-    keyUp(with: event)
-    return nil
-  }
-
-  private func localEventLeftMouseDown(_ event: NSEvent) -> NSEvent? {
-    guard let window, event.window != nil, window == event.window else { return event }
-    let location = convert(event.locationInWindow, from: nil)
-    guard hitTest(location) == self else { return event }
-    guard !NSApp.isActive || !window.isKeyWindow else { return event }
-    guard !focused else { return event }
-    window.makeFirstResponder(self)
-    return event
-  }
-
   func updateSurfaceSize() {
     resumeDeferredOcclusionIfNeeded()
     guard let surface else { return }
@@ -1234,1474 +816,5 @@ final class GhosttySurfaceView: NSView, Identifiable {
       }
       window.makeFirstResponder(view)
     }
-  }
-
-  override func performKeyEquivalent(with event: NSEvent) -> Bool {
-    guard event.type == .keyDown else { return false }
-    let isFontSizeShortcut = matchesFontSizeShortcut(event: event)
-    guard let surface else { return false }
-    guard
-      Self.hasKeyEquivalentFocusOwnership(
-        cachedFocused: focused,
-        isActualFirstResponder: window?.firstResponder === self
-      )
-    else { return false }
-
-    if UserCustomShortcutRegistry.shared.matches(event: event),
-      let menu = NSApp.mainMenu,
-      Self.mainMenuHasMatchingItem(for: event, in: menu),
-      menu.performKeyEquivalent(with: event)
-    {
-      return true
-    }
-
-    if let bindingFlags = bindingFlags(for: event, surface: surface) {
-      if shouldAttemptMenu(for: bindingFlags),
-        let menu = NSApp.mainMenu,
-        Self.mainMenuHasMatchingItem(for: event, in: menu),
-        menu.performKeyEquivalent(with: event)
-      {
-        return true
-      }
-      keyDown(with: event)
-      if isFontSizeShortcut {
-        onFontSizeShortcut?()
-      }
-      // Ghostty handled paste internally; broadcast the pasted text to followers.
-      if onCommittedText != nil,
-        event.modifierFlags.contains(.command),
-        event.charactersIgnoringModifiers == "v",
-        let text = NSPasteboard.general.string(forType: .string),
-        !text.isEmpty
-      {
-        onCommittedText?(text)
-      }
-      return true
-    }
-
-    guard let equivalent = equivalentKey(for: event) else { return false }
-
-    guard
-      let finalEvent = NSEvent.keyEvent(
-        with: .keyDown,
-        location: event.locationInWindow,
-        modifierFlags: event.modifierFlags,
-        timestamp: event.timestamp,
-        windowNumber: event.windowNumber,
-        context: nil,
-        characters: equivalent,
-        charactersIgnoringModifiers: equivalent,
-        isARepeat: event.isARepeat,
-        keyCode: event.keyCode
-      )
-    else {
-      return false
-    }
-    keyDown(with: finalEvent)
-    if isFontSizeShortcut {
-      onFontSizeShortcut?()
-    }
-    return true
-  }
-
-  private func matchesFontSizeShortcut(event: NSEvent) -> Bool {
-    matchesBindingShortcut(event: event, action: "reset_font_size")
-      || matchesBindingShortcut(event: event, action: "increase_font_size:1")
-      || matchesBindingShortcut(event: event, action: "decrease_font_size:1")
-  }
-
-  private func matchesBindingShortcut(event: NSEvent, action: String) -> Bool {
-    guard let shortcut = runtime.keyboardShortcut(for: action) else { return false }
-    let normalizedEventModifiers = normalizedModifiers(from: event.modifierFlags)
-    guard normalizedEventModifiers == shortcut.modifiers else { return false }
-    let eventKey = (event.charactersIgnoringModifiers ?? "").lowercased()
-    let shortcutKey = String(shortcut.key.character).lowercased()
-    return !eventKey.isEmpty && eventKey == shortcutKey
-  }
-
-  private func normalizedModifiers(from flags: NSEvent.ModifierFlags) -> SwiftUI.EventModifiers {
-    var normalized: SwiftUI.EventModifiers = []
-    if flags.contains(.command) { normalized.insert(.command) }
-    if flags.contains(.shift) { normalized.insert(.shift) }
-    if flags.contains(.option) { normalized.insert(.option) }
-    if flags.contains(.control) { normalized.insert(.control) }
-    return normalized
-  }
-
-  private func bindingFlags(
-    for event: NSEvent,
-    surface: ghostty_surface_t
-  ) -> ghostty_binding_flags_e? {
-    var key = ghosttyKeyEvent(
-      event,
-      action: GHOSTTY_ACTION_PRESS,
-      originalMods: event.modifierFlags,
-      translationMods: event.modifierFlags
-    )
-    var flags = ghostty_binding_flags_e(0)
-    let isBinding = (event.characters ?? "").withCString { ptr in
-      key.text = ptr
-      return ghostty_surface_key_is_binding(surface, key, &flags)
-    }
-    return isBinding ? flags : nil
-  }
-
-  private func equivalentKey(for event: NSEvent) -> String? {
-    switch event.charactersIgnoringModifiers {
-    case "\r":
-      guard event.modifierFlags.contains(.control) else { return nil }
-      return "\r"
-    case "/":
-      guard event.modifierFlags.contains(.control) else { return nil }
-      guard event.modifierFlags.isDisjoint(with: [.shift, .command, .option]) else { return nil }
-      return "_"
-    default:
-      if event.timestamp == 0 { return nil }
-      if !event.modifierFlags.contains(.command) && !event.modifierFlags.contains(.control) {
-        lastPerformKeyEvent = nil
-        return nil
-      }
-      if let lastPerformKeyEvent {
-        self.lastPerformKeyEvent = nil
-        if lastPerformKeyEvent == event.timestamp {
-          return event.characters ?? ""
-        }
-      }
-      lastPerformKeyEvent = event.timestamp
-      return nil
-    }
-  }
-
-  override func doCommand(by selector: Selector) {
-    if let lastPerformKeyEvent,
-      let current = NSApp.currentEvent,
-      lastPerformKeyEvent == current.timestamp
-    {
-      NSApp.sendEvent(current)
-      return
-    }
-    switch selector {
-    case #selector(moveToBeginningOfDocument(_:)):
-      performBindingAction("scroll_to_top")
-    case #selector(moveToEndOfDocument(_:)):
-      performBindingAction("scroll_to_bottom")
-    default:
-      break
-    }
-  }
-
-  override func menu(for event: NSEvent) -> NSMenu? {
-    switch event.type {
-    case .rightMouseDown:
-      break
-    case .leftMouseDown:
-      if !event.modifierFlags.contains(.control) {
-        return nil
-      }
-      guard let surface else { return nil }
-      if ghostty_surface_mouse_captured(surface) {
-        return nil
-      }
-      let mods = ghosttyMods(event.modifierFlags)
-      _ = ghostty_surface_mouse_button(surface, GHOSTTY_MOUSE_PRESS, GHOSTTY_MOUSE_RIGHT, mods)
-    default:
-      return nil
-    }
-
-    guard let surface else { return nil }
-    if ghostty_surface_mouse_captured(surface) {
-      return nil
-    }
-
-    let menu = NSMenu()
-    if ghostty_surface_has_selection(surface) {
-      menu.addItem(NSMenuItem(title: "Copy", action: #selector(copy(_:)), keyEquivalent: ""))
-    }
-    menu.addItem(NSMenuItem(title: "Paste", action: #selector(paste(_:)), keyEquivalent: ""))
-    menu.addItem(.separator())
-    menu.addItem(
-      menuItem(
-        title: "Split Right",
-        action: #selector(splitRight(_:)),
-        symbol: "rectangle.righthalf.inset.filled"
-      ))
-    menu.addItem(
-      menuItem(
-        title: "Split Left",
-        action: #selector(splitLeft(_:)),
-        symbol: "rectangle.leadinghalf.inset.filled"
-      ))
-    menu.addItem(
-      menuItem(
-        title: "Split Down",
-        action: #selector(splitDown(_:)),
-        symbol: "rectangle.bottomhalf.inset.filled"
-      ))
-    menu.addItem(
-      menuItem(
-        title: "Split Up",
-        action: #selector(splitUp(_:)),
-        symbol: "rectangle.tophalf.inset.filled"
-      ))
-    menu.addItem(.separator())
-    menu.addItem(
-      menuItem(
-        title: "Reset Terminal",
-        action: #selector(resetTerminal(_:)),
-        symbol: "arrow.trianglehead.2.clockwise"
-      ))
-    menu.addItem(.separator())
-    menu.addItem(
-      menuItem(
-        title: "Change Title...",
-        action: #selector(changeTitle(_:)),
-        symbol: "pencil.line"
-      ))
-    return menu
-  }
-
-  private func menuItem(title: String, action: Selector, symbol: String) -> NSMenuItem {
-    let item = NSMenuItem(title: title, action: action, keyEquivalent: "")
-    item.image = NSImage(systemSymbolName: symbol, accessibilityDescription: nil)
-    return item
-  }
-
-  @IBAction func splitRight(_ sender: Any?) {
-    _ = bridge.onSplitAction?(.newSplit(direction: .right))
-  }
-
-  @IBAction func splitLeft(_ sender: Any?) {
-    _ = bridge.onSplitAction?(.newSplit(direction: .left))
-  }
-
-  @IBAction func splitDown(_ sender: Any?) {
-    _ = bridge.onSplitAction?(.newSplit(direction: .down))
-  }
-
-  @IBAction func splitUp(_ sender: Any?) {
-    _ = bridge.onSplitAction?(.newSplit(direction: .top))
-  }
-
-  @IBAction func resetTerminal(_ sender: Any?) {
-    performBindingAction("reset")
-  }
-
-  @IBAction func changeTitle(_ sender: Any?) {
-    performBindingAction("prompt_surface_title")
-  }
-
-  static func hasKeyEquivalentFocusOwnership(
-    cachedFocused: Bool,
-    isActualFirstResponder: Bool
-  ) -> Bool {
-    cachedFocused && isActualFirstResponder
-  }
-
-  static func mainMenuHasMatchingItem(for event: NSEvent, in menu: NSMenu) -> Bool {
-    let eventEquivalents = normalizedEventKeyEquivalents(for: event)
-    guard !eventEquivalents.isEmpty else { return false }
-
-    for item in menu.items {
-      if let submenu = item.submenu, mainMenuHasMatchingItem(for: event, in: submenu) {
-        return true
-      }
-
-      guard !item.keyEquivalent.isEmpty else { continue }
-      guard
-        let itemEquivalent = normalizedKeyEquivalent(
-          key: item.keyEquivalent,
-          modifiers: item.keyEquivalentModifierMask
-        )
-      else { continue }
-      if eventEquivalents.contains(where: { $0 == itemEquivalent }) {
-        return true
-      }
-    }
-
-    return false
-  }
-
-  private static let shortcutMask: NSEvent.ModifierFlags = [.shift, .control, .option, .command]
-
-  private static let shiftedKeyEquivalentBases: [Character: Character] = [
-    "~": "`", "!": "1", "@": "2", "#": "3", "$": "4", "%": "5", "^": "6", "&": "7",
-    "*": "8", "(": "9", ")": "0", "_": "-", "+": "=", "{": "[", "}": "]", "|": "\\",
-    ":": ";", "\"": "'", "<": ",", ">": ".", "?": "/",
-  ]
-
-  private struct KeyEquivalentSignature: Equatable {
-    var key: String
-    var modifiers: NSEvent.ModifierFlags
-  }
-
-  private static func normalizedEventKeyEquivalents(for event: NSEvent) -> [KeyEquivalentSignature] {
-    let eventModifiers = event.modifierFlags.intersection(shortcutMask)
-    return [event.charactersIgnoringModifiers, event.characters]
-      .compactMap { characters in
-        characters.flatMap { normalizedKeyEquivalent(key: $0, modifiers: eventModifiers) }
-      }
-      .reduce(into: []) { result, equivalent in
-        if !result.contains(equivalent) {
-          result.append(equivalent)
-        }
-      }
-  }
-
-  private static func normalizedKeyEquivalent(
-    key: String,
-    modifiers: NSEvent.ModifierFlags
-  ) -> KeyEquivalentSignature? {
-    guard !key.isEmpty else { return nil }
-
-    var normalizedKey = key.lowercased()
-    var normalizedModifiers = modifiers.intersection(shortcutMask)
-
-    if key.count == 1, let character = key.first {
-      if let base = shiftedKeyEquivalentBases[character] {
-        normalizedKey = String(base)
-        normalizedModifiers.insert(.shift)
-      } else if normalizedKey != key {
-        normalizedModifiers.insert(.shift)
-      }
-    }
-
-    return KeyEquivalentSignature(key: normalizedKey, modifiers: normalizedModifiers)
-  }
-
-  private func shouldAttemptMenu(for flags: ghostty_binding_flags_e) -> Bool {
-    if bridge.state.keySequenceActive == true { return false }
-    if bridge.state.keyTableDepth > 0 { return false }
-    let raw = flags.rawValue
-    let isAll = (raw & GHOSTTY_BINDING_FLAGS_ALL.rawValue) != 0
-    let isPerformable = (raw & GHOSTTY_BINDING_FLAGS_PERFORMABLE.rawValue) != 0
-    let isConsumed = (raw & GHOSTTY_BINDING_FLAGS_CONSUMED.rawValue) != 0
-    return !isAll && !isPerformable && isConsumed
-  }
-
-  @IBAction func copy(_ sender: Any?) {
-    performBindingAction("copy_to_clipboard")
-  }
-
-  @IBAction func paste(_ sender: Any?) {
-    performBindingAction("paste_from_clipboard")
-    if let text = NSPasteboard.general.string(forType: .string), !text.isEmpty {
-      onCommittedText?(text)
-    }
-  }
-
-  @IBAction func pasteSelection(_ sender: Any?) {
-    performBindingAction("paste_from_selection")
-  }
-
-  @IBAction override func selectAll(_ sender: Any?) {
-    performBindingAction("select_all")
-  }
-
-  @discardableResult
-  private func sendKey(
-    action: ghostty_input_action_e,
-    event: NSEvent,
-    translationEvent: NSEvent? = nil,
-    translationMods: NSEvent.ModifierFlags? = nil,
-    text: String? = nil,
-    composing: Bool = false
-  ) -> Bool {
-    guard let surface else { return false }
-    let resolvedEvent: NSEvent
-    let resolvedMods: NSEvent.ModifierFlags
-    if let translationEvent, let translationMods {
-      resolvedEvent = translationEvent
-      resolvedMods = translationMods
-    } else {
-      (resolvedEvent, resolvedMods) = translationState(event, surface: surface)
-    }
-    var key = ghosttyKeyEvent(
-      resolvedEvent,
-      action: action,
-      originalMods: event.modifierFlags,
-      translationMods: resolvedMods,
-      composing: composing
-    )
-    let finalText = text ?? ghosttyCharacters(resolvedEvent)
-    if let finalText, !finalText.isEmpty,
-      let codepoint = finalText.utf8.first, codepoint >= 0x20
-    {
-      return finalText.withCString { ptr in
-        key.text = ptr
-        return ghostty_surface_key(surface, key)
-      }
-    }
-    key.text = nil
-    return ghostty_surface_key(surface, key)
-  }
-
-  func performBindingAction(_ action: String) {
-    guard let surface else { return }
-    _ = action.withCString { ptr in
-      ghostty_surface_binding_action(surface, ptr, UInt(action.lengthOfBytes(using: .utf8)))
-    }
-  }
-
-  private func translationState(_ event: NSEvent, surface: ghostty_surface_t) -> (
-    NSEvent, NSEvent.ModifierFlags
-  ) {
-    let translatedModsGhostty = ghostty_surface_key_translation_mods(
-      surface, ghosttyMods(event.modifierFlags))
-    let translatedMods = appKitMods(translatedModsGhostty)
-    var resolved = event.modifierFlags
-    for flag in [NSEvent.ModifierFlags.shift, .control, .option, .command] {
-      if translatedMods.contains(flag) {
-        resolved.insert(flag)
-      } else {
-        resolved.remove(flag)
-      }
-    }
-    if resolved == event.modifierFlags {
-      return (event, resolved)
-    }
-    let translatedEvent =
-      NSEvent.keyEvent(
-        with: event.type,
-        location: event.locationInWindow,
-        modifierFlags: resolved,
-        timestamp: event.timestamp,
-        windowNumber: event.windowNumber,
-        context: nil,
-        characters: event.characters(byApplyingModifiers: resolved) ?? "",
-        charactersIgnoringModifiers: event.charactersIgnoringModifiers ?? "",
-        isARepeat: event.isARepeat,
-        keyCode: event.keyCode
-      ) ?? event
-    return (translatedEvent, resolved)
-  }
-
-  private func ghosttyKeyEvent(
-    _ event: NSEvent,
-    action: ghostty_input_action_e,
-    originalMods: NSEvent.ModifierFlags,
-    translationMods: NSEvent.ModifierFlags,
-    composing: Bool = false
-  ) -> ghostty_input_key_s {
-    var keyEvent: ghostty_input_key_s = .init()
-    keyEvent.action = action
-    keyEvent.keycode = UInt32(event.keyCode)
-    keyEvent.text = nil
-    keyEvent.composing = composing
-    keyEvent.mods = ghosttyMods(originalMods)
-    keyEvent.consumed_mods = ghosttyMods(translationMods.subtracting([.control, .command]))
-    keyEvent.unshifted_codepoint = 0
-    if event.type == .keyDown || event.type == .keyUp {
-      if let chars = event.characters(byApplyingModifiers: []),
-        let codepoint = chars.unicodeScalars.first
-      {
-        keyEvent.unshifted_codepoint = codepoint.value
-      }
-    }
-    return keyEvent
-  }
-
-  private func suppressKeyboardLayoutChangeKeyUp(_ event: NSEvent) -> Bool {
-    guard let suppression = keyboardLayoutChangeKeyUpSuppression else { return false }
-    if suppression.isExpired(at: event.timestamp) {
-      keyboardLayoutChangeKeyUpSuppression = nil
-      return false
-    }
-    if suppression.suppresses(keyCode: event.keyCode, timestamp: event.timestamp) {
-      keyboardLayoutChangeKeyUpSuppression = nil
-      return true
-    }
-    return false
-  }
-
-  private func ghosttyCharacters(_ event: NSEvent) -> String? {
-    GhosttyEventText.characters(for: event)
-  }
-
-  private func syncPreedit(clearIfNeeded: Bool = true) {
-    guard let surface else { return }
-    if markedText.length > 0 {
-      let str = markedText.string
-      let len = str.utf8CString.count
-      if len > 0 {
-        markedText.string.withCString { ptr in
-          ghostty_surface_preedit(surface, ptr, UInt(len - 1))
-        }
-      }
-    } else if clearIfNeeded {
-      ghostty_surface_preedit(surface, nil, 0)
-    }
-  }
-
-  private func scrollMods(for event: NSEvent) -> ghostty_input_scroll_mods_t {
-    var value: Int32 = 0
-    if event.hasPreciseScrollingDeltas {
-      value |= 0b0000_0001
-    }
-    let momentum: Int32
-    switch event.momentumPhase {
-    case .began:
-      momentum = 1
-    case .stationary:
-      momentum = 2
-    case .changed:
-      momentum = 3
-    case .ended:
-      momentum = 4
-    case .cancelled:
-      momentum = 5
-    case .mayBegin:
-      momentum = 6
-    default:
-      momentum = 0
-    }
-    value |= (momentum << 1)
-    return ghostty_input_scroll_mods_t(value)
-  }
-
-  private func keyboardLayoutId() -> String? {
-    guard let source = TISCopyCurrentKeyboardLayoutInputSource()?.takeRetainedValue() else {
-      return nil
-    }
-    guard let raw = TISGetInputSourceProperty(source, kTISPropertyInputSourceID) else {
-      return nil
-    }
-    let value = Unmanaged<CFString>.fromOpaque(raw).takeUnretainedValue()
-    return value as String
-  }
-
-  private func sendMousePosition(_ event: NSEvent) {
-    guard let surface else { return }
-    let point = convert(event.locationInWindow, from: nil)
-    let yPosition = bounds.height - point.y
-    let mods = ghosttyMods(event.modifierFlags)
-    ghostty_surface_mouse_pos(surface, point.x, yPosition, mods)
-  }
-
-  private func sendMouseButton(
-    _ event: NSEvent,
-    state: ghostty_input_mouse_state_e,
-    button: ghostty_input_mouse_button_e
-  ) {
-    guard let surface else { return }
-    let mods = ghosttyMods(event.modifierFlags)
-    ghostty_surface_mouse_button(surface, state, button, mods)
-  }
-
-  private func ghosttyMods(_ flags: NSEvent.ModifierFlags) -> ghostty_input_mods_e {
-    var mods: UInt32 = GHOSTTY_MODS_NONE.rawValue
-    if flags.contains(.shift) { mods |= GHOSTTY_MODS_SHIFT.rawValue }
-    if flags.contains(.control) { mods |= GHOSTTY_MODS_CTRL.rawValue }
-    if flags.contains(.option) { mods |= GHOSTTY_MODS_ALT.rawValue }
-    if flags.contains(.command) { mods |= GHOSTTY_MODS_SUPER.rawValue }
-    if flags.contains(.capsLock) { mods |= GHOSTTY_MODS_CAPS.rawValue }
-    let rawFlags = flags.rawValue
-    if (rawFlags & UInt(NX_DEVICERSHIFTKEYMASK)) != 0 { mods |= GHOSTTY_MODS_SHIFT_RIGHT.rawValue }
-    if (rawFlags & UInt(NX_DEVICERCTLKEYMASK)) != 0 { mods |= GHOSTTY_MODS_CTRL_RIGHT.rawValue }
-    if (rawFlags & UInt(NX_DEVICERALTKEYMASK)) != 0 { mods |= GHOSTTY_MODS_ALT_RIGHT.rawValue }
-    if (rawFlags & UInt(NX_DEVICERCMDKEYMASK)) != 0 { mods |= GHOSTTY_MODS_SUPER_RIGHT.rawValue }
-    return ghostty_input_mods_e(mods)
-  }
-
-  private func appKitMods(_ mods: ghostty_input_mods_e) -> NSEvent.ModifierFlags {
-    var flags: NSEvent.ModifierFlags = []
-    if (mods.rawValue & GHOSTTY_MODS_SHIFT.rawValue) != 0 { flags.insert(.shift) }
-    if (mods.rawValue & GHOSTTY_MODS_CTRL.rawValue) != 0 { flags.insert(.control) }
-    if (mods.rawValue & GHOSTTY_MODS_ALT.rawValue) != 0 { flags.insert(.option) }
-    if (mods.rawValue & GHOSTTY_MODS_SUPER.rawValue) != 0 { flags.insert(.command) }
-    if (mods.rawValue & GHOSTTY_MODS_CAPS.rawValue) != 0 { flags.insert(.capsLock) }
-    return flags
-  }
-
-}
-
-extension GhosttySurfaceView {
-  override func draggingEntered(_ sender: any NSDraggingInfo) -> NSDragOperation {
-    guard let types = sender.draggingPasteboard.types else { return [] }
-    if Set(types).isDisjoint(with: Self.dropTypes) {
-      return []
-    }
-    return .copy
-  }
-
-  override func performDragOperation(_ sender: any NSDraggingInfo) -> Bool {
-    let pasteboard = sender.draggingPasteboard
-    let content: String?
-    if let url = pasteboard.string(forType: .URL) {
-      content = NSPasteboard.ghosttyEscape(url)
-    } else if let urls = pasteboard.readObjects(forClasses: [NSURL.self]) as? [URL],
-      !urls.isEmpty
-    {
-      content = urls.map { NSPasteboard.ghosttyEscape($0.path) }.joined(separator: " ")
-    } else if let str = pasteboard.string(forType: .string) {
-      content = str
-    } else {
-      content = nil
-    }
-
-    guard let content else { return false }
-    Task { @MainActor in
-      self.insertText(content, replacementRange: NSRange(location: 0, length: 0))
-    }
-    return true
-  }
-}
-
-extension GhosttySurfaceView: NSTextInputClient {
-  func hasMarkedText() -> Bool {
-    markedText.length > 0
-  }
-
-  func markedRange() -> NSRange {
-    guard markedText.length > 0 else { return NSRange() }
-    return NSRange(location: 0, length: markedText.length)
-  }
-
-  func selectedRange() -> NSRange {
-    guard let surface else { return NSRange() }
-    var text = ghostty_text_s()
-    guard ghostty_surface_read_selection(surface, &text) else { return NSRange() }
-    defer { ghostty_surface_free_text(surface, &text) }
-    return NSRange(location: Int(text.offset_start), length: Int(text.offset_len))
-  }
-
-  func setMarkedText(_ string: Any, selectedRange: NSRange, replacementRange: NSRange) {
-    switch string {
-    case let attributedText as NSAttributedString:
-      markedText = NSMutableAttributedString(attributedString: attributedText)
-    case let stringValue as String:
-      markedText = NSMutableAttributedString(string: stringValue)
-    default:
-      return
-    }
-    if keyTextAccumulator == nil {
-      syncPreedit()
-    }
-  }
-
-  func unmarkText() {
-    if markedText.length > 0 {
-      markedText.mutableString.setString("")
-      syncPreedit()
-    }
-  }
-
-  func validAttributesForMarkedText() -> [NSAttributedString.Key] {
-    []
-  }
-
-  func attributedSubstring(
-    forProposedRange range: NSRange,
-    actualRange: NSRangePointer?
-  ) -> NSAttributedString? {
-    guard let surface else { return nil }
-    guard range.length > 0 else { return nil }
-    var text = ghostty_text_s()
-    guard ghostty_surface_read_selection(surface, &text) else { return nil }
-    defer { ghostty_surface_free_text(surface, &text) }
-    var attributes: [NSAttributedString.Key: Any] = [:]
-    if let fontRaw = ghostty_surface_quicklook_font(surface) {
-      let font = Unmanaged<CTFont>.fromOpaque(fontRaw)
-      attributes[.font] = font.takeUnretainedValue()
-      font.release()
-    }
-    return NSAttributedString(string: Self.string(from: text), attributes: attributes)
-  }
-
-  func characterIndex(for point: NSPoint) -> Int {
-    0
-  }
-
-  func firstRect(forCharacterRange range: NSRange, actualRange: NSRangePointer?) -> NSRect {
-    guard let surface else {
-      return NSRect(x: frame.origin.x, y: frame.origin.y, width: 0, height: 0)
-    }
-    var caretX: Double = 0
-    var caretY: Double = 0
-    var width: Double = cellSize.width
-    var height: Double = cellSize.height
-    if range.length > 0, range != selectedRange() {
-      var text = ghostty_text_s()
-      if ghostty_surface_read_selection(surface, &text) {
-        caretX = text.tl_px_x - 2
-        caretY = text.tl_px_y + 2
-        ghostty_surface_free_text(surface, &text)
-      } else {
-        ghostty_surface_ime_point(surface, &caretX, &caretY, &width, &height)
-      }
-    } else {
-      ghostty_surface_ime_point(surface, &caretX, &caretY, &width, &height)
-    }
-    if range.length == 0, width > 0 {
-      width = 0
-      caretX += cellSize.width * Double(range.location + range.length)
-    }
-    let viewRect = NSRect(
-      x: caretX,
-      y: frame.size.height - caretY,
-      width: width,
-      height: max(height, cellSize.height)
-    )
-    let winRect = convert(viewRect, to: nil)
-    guard let window else { return winRect }
-    return window.convertToScreen(winRect)
-  }
-
-  func insertText(_ string: Any, replacementRange _: NSRange) {
-    guard NSApp.currentEvent != nil else { return }
-    guard surface != nil else { return }
-    var chars = ""
-    switch string {
-    case let attributedText as NSAttributedString:
-      chars = attributedText.string
-    case let stringValue as String:
-      chars = stringValue
-    default:
-      return
-    }
-    unmarkText()
-    if var acc = keyTextAccumulator {
-      acc.append(chars)
-      keyTextAccumulator = acc
-      onCommittedText?(chars)
-      return
-    }
-    insertCommittedTextForBroadcast(chars)
-    onCommittedText?(chars)
-  }
-
-  func insertCommittedTextForBroadcast(_ text: String) {
-    guard let surface else { return }
-    guard !text.isEmpty else { return }
-    unmarkText()
-    let len = text.utf8CString.count
-    guard len > 0 else { return }
-    text.withCString { ptr in
-      ghostty_surface_text(surface, ptr, UInt(len - 1))
-    }
-  }
-
-  @discardableResult
-  func applyMirroredKeyForBroadcast(_ key: MirroredTerminalKey) -> Bool {
-    let windowNumber = window?.windowNumber ?? 0
-    guard let keyDownEvent = key.keyDownEvent(windowNumber: windowNumber) else {
-      return false
-    }
-    keyDown(with: keyDownEvent)
-    if !key.isRepeat,
-      let keyUpEvent = key.keyUpEvent(windowNumber: windowNumber)
-    {
-      keyUp(with: keyUpEvent)
-    }
-    return true
-  }
-
-  @discardableResult
-  func submitLine() -> Bool {
-    let timestamp = ProcessInfo.processInfo.systemUptime
-    let windowNumber = window?.windowNumber ?? 0
-    guard
-      let keyDownEvent = NSEvent.keyEvent(
-        with: .keyDown,
-        location: .zero,
-        modifierFlags: [],
-        timestamp: timestamp,
-        windowNumber: windowNumber,
-        context: nil,
-        characters: "\r",
-        charactersIgnoringModifiers: "\r",
-        isARepeat: false,
-        keyCode: 36
-      ),
-      let keyUpEvent = NSEvent.keyEvent(
-        with: .keyUp,
-        location: .zero,
-        modifierFlags: [],
-        timestamp: timestamp,
-        windowNumber: windowNumber,
-        context: nil,
-        characters: "\r",
-        charactersIgnoringModifiers: "\r",
-        isARepeat: false,
-        keyCode: 36
-      )
-    else {
-      return false
-    }
-    keyDown(with: keyDownEvent)
-    keyUp(with: keyUpEvent)
-    return true
-  }
-
-  // MARK: - CLI key token delivery
-
-  /// Send a single key event for a canonical CLI token (e.g. "enter", "ctrl-c", "pageup").
-  /// Creates synthetic keyDown/keyUp NSEvents matching the token spec.
-  @discardableResult
-  func sendCLIKeyToken(_ token: String) -> Bool {
-    guard surface != nil else { return false }
-    guard let spec = CLIKeySpec.from(token: token) else { return false }
-    let timestamp = ProcessInfo.processInfo.systemUptime
-    let windowNumber = window?.windowNumber ?? 0
-    guard
-      let keyDownEvent = NSEvent.keyEvent(
-        with: .keyDown,
-        location: .zero,
-        modifierFlags: spec.modifiers,
-        timestamp: timestamp,
-        windowNumber: windowNumber,
-        context: nil,
-        characters: spec.characters,
-        charactersIgnoringModifiers: spec.charactersIgnoringModifiers,
-        isARepeat: false,
-        keyCode: spec.keyCode
-      ),
-      let keyUpEvent = NSEvent.keyEvent(
-        with: .keyUp,
-        location: .zero,
-        modifierFlags: spec.modifiers,
-        timestamp: timestamp,
-        windowNumber: windowNumber,
-        context: nil,
-        characters: spec.characters,
-        charactersIgnoringModifiers: spec.charactersIgnoringModifiers,
-        isARepeat: false,
-        keyCode: spec.keyCode
-      )
-    else {
-      return false
-    }
-    keyDown(with: keyDownEvent)
-    keyUp(with: keyUpEvent)
-    return true
-  }
-}
-
-// MARK: - CLI Key Spec
-
-/// Maps canonical CLI key tokens to macOS NSEvent parameters.
-struct CLIKeySpec {
-  let keyCode: UInt16
-  let characters: String
-  let charactersIgnoringModifiers: String
-  let modifiers: NSEvent.ModifierFlags
-
-  static func from(token: String) -> CLIKeySpec? {
-    guard let descriptor = KeyTokens.descriptor(for: token),
-      let base = baseSpec(for: descriptor.baseToken)
-    else {
-      return nil
-    }
-
-    var modifierFlags = eventModifiers(from: descriptor.modifiers)
-    if base.usesFunctionModifier {
-      modifierFlags.insert(.function)
-    }
-
-    let charactersIgnoringModifiers = base.charactersIgnoringModifiers
-    let characters: String
-    if usesControlCharacter(for: descriptor.modifiers),
-      let controlCharacter = controlCharacter(for: descriptor.baseToken)
-    {
-      characters = controlCharacter
-    } else if descriptor.modifiers.contains(.shift) {
-      characters = shiftedCharacters(for: descriptor.baseToken) ?? charactersIgnoringModifiers
-    } else {
-      characters = charactersIgnoringModifiers
-    }
-
-    return CLIKeySpec(
-      keyCode: base.keyCode,
-      characters: characters,
-      charactersIgnoringModifiers: charactersIgnoringModifiers,
-      modifiers: modifierFlags
-    )
-  }
-
-  private struct BaseSpec {
-    let keyCode: UInt16
-    let charactersIgnoringModifiers: String
-    let usesFunctionModifier: Bool
-  }
-
-  private static let namedBaseSpecs: [String: BaseSpec] = [
-    "enter": BaseSpec(
-      keyCode: UInt16(kVK_Return),
-      charactersIgnoringModifiers: "\r",
-      usesFunctionModifier: false
-    ),
-    "esc": BaseSpec(
-      keyCode: UInt16(kVK_Escape),
-      charactersIgnoringModifiers: "\u{1B}",
-      usesFunctionModifier: false
-    ),
-    "tab": BaseSpec(
-      keyCode: UInt16(kVK_Tab),
-      charactersIgnoringModifiers: "\t",
-      usesFunctionModifier: false
-    ),
-    "backspace": BaseSpec(
-      keyCode: UInt16(kVK_Delete),
-      charactersIgnoringModifiers: "\u{7F}",
-      usesFunctionModifier: false
-    ),
-    "delete-forward": BaseSpec(
-      keyCode: UInt16(kVK_ForwardDelete),
-      charactersIgnoringModifiers: String(UnicodeScalar(NSDeleteFunctionKey)!),
-      usesFunctionModifier: true
-    ),
-    "insert": BaseSpec(
-      keyCode: UInt16(kVK_Help),
-      charactersIgnoringModifiers: String(UnicodeScalar(NSInsertFunctionKey)!),
-      usesFunctionModifier: true
-    ),
-    "up": BaseSpec(
-      keyCode: UInt16(kVK_UpArrow),
-      charactersIgnoringModifiers: String(UnicodeScalar(NSUpArrowFunctionKey)!),
-      usesFunctionModifier: true
-    ),
-    "down": BaseSpec(
-      keyCode: UInt16(kVK_DownArrow),
-      charactersIgnoringModifiers: String(UnicodeScalar(NSDownArrowFunctionKey)!),
-      usesFunctionModifier: true
-    ),
-    "left": BaseSpec(
-      keyCode: UInt16(kVK_LeftArrow),
-      charactersIgnoringModifiers: String(UnicodeScalar(NSLeftArrowFunctionKey)!),
-      usesFunctionModifier: true
-    ),
-    "right": BaseSpec(
-      keyCode: UInt16(kVK_RightArrow),
-      charactersIgnoringModifiers: String(UnicodeScalar(NSRightArrowFunctionKey)!),
-      usesFunctionModifier: true
-    ),
-    "pageup": BaseSpec(
-      keyCode: UInt16(kVK_PageUp),
-      charactersIgnoringModifiers: String(UnicodeScalar(NSPageUpFunctionKey)!),
-      usesFunctionModifier: true
-    ),
-    "pagedown": BaseSpec(
-      keyCode: UInt16(kVK_PageDown),
-      charactersIgnoringModifiers: String(UnicodeScalar(NSPageDownFunctionKey)!),
-      usesFunctionModifier: true
-    ),
-    "home": BaseSpec(
-      keyCode: UInt16(kVK_Home),
-      charactersIgnoringModifiers: String(UnicodeScalar(NSHomeFunctionKey)!),
-      usesFunctionModifier: true
-    ),
-    "end": BaseSpec(
-      keyCode: UInt16(kVK_End),
-      charactersIgnoringModifiers: String(UnicodeScalar(NSEndFunctionKey)!),
-      usesFunctionModifier: true
-    ),
-  ]
-
-  private static let functionKeyMap: [String: (UInt16, Int)] = [
-    "f1": (UInt16(kVK_F1), NSF1FunctionKey),
-    "f2": (UInt16(kVK_F2), NSF2FunctionKey),
-    "f3": (UInt16(kVK_F3), NSF3FunctionKey),
-    "f4": (UInt16(kVK_F4), NSF4FunctionKey),
-    "f5": (UInt16(kVK_F5), NSF5FunctionKey),
-    "f6": (UInt16(kVK_F6), NSF6FunctionKey),
-    "f7": (UInt16(kVK_F7), NSF7FunctionKey),
-    "f8": (UInt16(kVK_F8), NSF8FunctionKey),
-    "f9": (UInt16(kVK_F9), NSF9FunctionKey),
-    "f10": (UInt16(kVK_F10), NSF10FunctionKey),
-    "f11": (UInt16(kVK_F11), NSF11FunctionKey),
-    "f12": (UInt16(kVK_F12), NSF12FunctionKey),
-  ]
-
-  private static let printableKeyCodes: [String: UInt16] = [
-    "a": UInt16(kVK_ANSI_A),
-    "b": UInt16(kVK_ANSI_B),
-    "c": UInt16(kVK_ANSI_C),
-    "d": UInt16(kVK_ANSI_D),
-    "e": UInt16(kVK_ANSI_E),
-    "f": UInt16(kVK_ANSI_F),
-    "g": UInt16(kVK_ANSI_G),
-    "h": UInt16(kVK_ANSI_H),
-    "i": UInt16(kVK_ANSI_I),
-    "j": UInt16(kVK_ANSI_J),
-    "k": UInt16(kVK_ANSI_K),
-    "l": UInt16(kVK_ANSI_L),
-    "m": UInt16(kVK_ANSI_M),
-    "n": UInt16(kVK_ANSI_N),
-    "o": UInt16(kVK_ANSI_O),
-    "p": UInt16(kVK_ANSI_P),
-    "q": UInt16(kVK_ANSI_Q),
-    "r": UInt16(kVK_ANSI_R),
-    "s": UInt16(kVK_ANSI_S),
-    "t": UInt16(kVK_ANSI_T),
-    "u": UInt16(kVK_ANSI_U),
-    "v": UInt16(kVK_ANSI_V),
-    "w": UInt16(kVK_ANSI_W),
-    "x": UInt16(kVK_ANSI_X),
-    "y": UInt16(kVK_ANSI_Y),
-    "z": UInt16(kVK_ANSI_Z),
-    "0": UInt16(kVK_ANSI_0),
-    "1": UInt16(kVK_ANSI_1),
-    "2": UInt16(kVK_ANSI_2),
-    "3": UInt16(kVK_ANSI_3),
-    "4": UInt16(kVK_ANSI_4),
-    "5": UInt16(kVK_ANSI_5),
-    "6": UInt16(kVK_ANSI_6),
-    "7": UInt16(kVK_ANSI_7),
-    "8": UInt16(kVK_ANSI_8),
-    "9": UInt16(kVK_ANSI_9),
-    "space": UInt16(kVK_Space),
-    "minus": UInt16(kVK_ANSI_Minus),
-    "equal": UInt16(kVK_ANSI_Equal),
-    "comma": UInt16(kVK_ANSI_Comma),
-    "period": UInt16(kVK_ANSI_Period),
-    "slash": UInt16(kVK_ANSI_Slash),
-    "backslash": UInt16(kVK_ANSI_Backslash),
-    "semicolon": UInt16(kVK_ANSI_Semicolon),
-    "quote": UInt16(kVK_ANSI_Quote),
-    "grave": UInt16(kVK_ANSI_Grave),
-    "left-bracket": UInt16(kVK_ANSI_LeftBracket),
-    "right-bracket": UInt16(kVK_ANSI_RightBracket),
-  ]
-
-  private static let shiftedCharacterMap: [String: String] = [
-    "1": "!",
-    "2": "@",
-    "3": "#",
-    "4": "$",
-    "5": "%",
-    "6": "^",
-    "7": "&",
-    "8": "*",
-    "9": "(",
-    "0": ")",
-    "minus": "_",
-    "equal": "+",
-    "comma": "<",
-    "period": ">",
-    "slash": "?",
-    "backslash": "|",
-    "semicolon": ":",
-    "quote": "\"",
-    "grave": "~",
-    "left-bracket": "{",
-    "right-bracket": "}",
-  ]
-
-  private static func eventModifiers(from modifiers: [KeyModifier]) -> NSEvent.ModifierFlags {
-    modifiers.reduce(into: NSEvent.ModifierFlags()) { result, modifier in
-      switch modifier {
-      case .cmd: result.insert(.command)
-      case .shift: result.insert(.shift)
-      case .opt: result.insert(.option)
-      case .ctrl: result.insert(.control)
-      }
-    }
-  }
-
-  private static func baseSpec(for token: String) -> BaseSpec? {
-    if let base = namedBaseSpecs[token] { return base }
-    if let (keyCode, scalar) = functionKeyMap[token] {
-      return BaseSpec(
-        keyCode: keyCode,
-        charactersIgnoringModifiers: String(UnicodeScalar(scalar)!),
-        usesFunctionModifier: true
-      )
-    }
-    return printableBaseSpec(for: token)
-  }
-
-  private static func printableBaseSpec(for token: String) -> BaseSpec? {
-    guard let character = printableCharacter(for: token),
-      let keyCode = printableKeyCode(for: token)
-    else { return nil }
-
-    return BaseSpec(
-      keyCode: keyCode,
-      charactersIgnoringModifiers: String(character),
-      usesFunctionModifier: false
-    )
-  }
-
-  private static func printableCharacter(for token: String) -> Character? {
-    switch token {
-    case "space": return " "
-    case "minus": return "-"
-    case "equal": return "="
-    case "comma": return ","
-    case "period": return "."
-    case "slash": return "/"
-    case "backslash": return "\\"
-    case "semicolon": return ";"
-    case "quote": return "'"
-    case "grave": return "`"
-    case "left-bracket": return "["
-    case "right-bracket": return "]"
-    default:
-      guard token.count == 1 else { return nil }
-      return token.first
-    }
-  }
-
-  private static func printableKeyCode(for token: String) -> UInt16? {
-    printableKeyCodes[token]
-  }
-
-  private static func shiftedCharacters(for token: String) -> String? {
-    if token.count == 1, let character = token.first, character.isLetter {
-      return String(character).uppercased()
-    }
-
-    return shiftedCharacterMap[token]
-  }
-
-  private static func usesControlCharacter(for modifiers: [KeyModifier]) -> Bool {
-    let modifierSet = Set(modifiers)
-    return modifierSet.contains(.ctrl) && modifierSet.isSubset(of: [.ctrl, .shift])
-  }
-
-  private static func controlCharacter(for token: String) -> String? {
-    if token.count == 1, let scalar = token.unicodeScalars.first, scalar.properties.isAlphabetic {
-      return String(UnicodeScalar(scalar.value & 0x1F)!)
-    }
-
-    switch token {
-    case "left-bracket": return String(UnicodeScalar(27)!)
-    case "backslash": return String(UnicodeScalar(28)!)
-    case "right-bracket": return String(UnicodeScalar(29)!)
-    case "6": return String(UnicodeScalar(30)!)
-    case "minus": return String(UnicodeScalar(31)!)
-    default: return nil
-    }
-  }
-}
-
-extension GhosttySurfaceView: NSServicesMenuRequestor {
-  override func validRequestor(
-    forSendType sendType: NSPasteboard.PasteboardType?,
-    returnType: NSPasteboard.PasteboardType?
-  ) -> Any? {
-    let receivable: [NSPasteboard.PasteboardType] = [.string, .init("public.utf8-plain-text")]
-    let sendable = receivable
-    let sendableRequiresSelection = sendable
-
-    if (returnType == nil || receivable.contains(returnType!))
-      && (sendType == nil || sendable.contains(sendType!))
-    {
-      if let sendType, sendableRequiresSelection.contains(sendType) {
-        if surface == nil || !ghostty_surface_has_selection(surface) {
-          return super.validRequestor(forSendType: sendType, returnType: returnType)
-        }
-      }
-      return self
-    }
-    return super.validRequestor(forSendType: sendType, returnType: returnType)
-  }
-
-  func writeSelection(to pboard: NSPasteboard, types: [NSPasteboard.PasteboardType]) -> Bool {
-    guard let surface else { return false }
-    var text = ghostty_text_s()
-    guard ghostty_surface_read_selection(surface, &text) else { return false }
-    defer { ghostty_surface_free_text(surface, &text) }
-    pboard.declareTypes([.string], owner: nil)
-    pboard.setString(Self.string(from: text), forType: .string)
-    return true
-  }
-
-  func readSelection(from pboard: NSPasteboard) -> Bool {
-    guard let str = pboard.getOpinionatedStringContents() else { return false }
-    let len = str.utf8CString.count
-    if len == 0 { return true }
-    str.withCString { ptr in
-      ghostty_surface_text(surface, ptr, UInt(len - 1))
-    }
-    return true
-  }
-}
-
-final class GhosttySurfaceScrollView: NSView {
-  enum HostKind: String {
-    case terminal
-    case canvas
-  }
-
-  private struct ScrollbarState {
-    let total: UInt64
-    let offset: UInt64
-    let length: UInt64
-  }
-
-  private let scrollView: NSScrollView
-  private let documentView: NSView
-  private let surfaceView: GhosttySurfaceView
-  let hostKind: HostKind
-  private let debugID = String(UUID().uuidString.prefix(8))
-  var debugIdentifier: String {
-    debugID
-  }
-  private var observers: [NSObjectProtocol] = []
-
-  private var isLiveScrolling = false
-  private var isProgrammaticScrollChange = false
-  private var isUserScrolledBack = false
-  private var lastSentRow: Int?
-  private var scrollbar: ScrollbarState?
-
-  /// When set, the surface renders at this fixed size regardless of the hosting
-  /// view's bounds. Used in canvas mode to prevent `.scaleEffect()` from causing
-  /// terminal reflow.
-  var pinnedSize: CGSize?
-
-  init(surfaceView: GhosttySurfaceView, hostKind: HostKind) {
-    self.surfaceView = surfaceView
-    self.hostKind = hostKind
-    scrollView = NSScrollView()
-    scrollView.hasHorizontalScroller = false
-    scrollView.autohidesScrollers = false
-    scrollView.usesPredominantAxisScrolling = true
-    scrollView.scrollerStyle = .overlay
-    scrollView.drawsBackground = false
-    scrollView.contentView.clipsToBounds = false
-    documentView = NSView(frame: .zero)
-    scrollView.documentView = documentView
-    documentView.addSubview(surfaceView)
-    super.init(frame: .zero)
-    addSubview(scrollView)
-    surfaceView.scrollWrapper = self
-    refreshAppearance()
-
-    scrollView.contentView.postsBoundsChangedNotifications = true
-    observers.append(
-      NotificationCenter.default.addObserver(
-        forName: NSView.boundsDidChangeNotification,
-        object: scrollView.contentView,
-        queue: .main
-      ) { [weak self] _ in
-        MainActor.assumeIsolated {
-          self?.handleScrollChange()
-        }
-      })
-
-    observers.append(
-      NotificationCenter.default.addObserver(
-        forName: NSScrollView.willStartLiveScrollNotification,
-        object: scrollView,
-        queue: .main
-      ) { [weak self] _ in
-        MainActor.assumeIsolated {
-          self?.isLiveScrolling = true
-        }
-      })
-
-    observers.append(
-      NotificationCenter.default.addObserver(
-        forName: NSScrollView.didEndLiveScrollNotification,
-        object: scrollView,
-        queue: .main
-      ) { [weak self] _ in
-        MainActor.assumeIsolated {
-          self?.isLiveScrolling = false
-          self?.updateScrollBackState()
-        }
-      })
-
-    observers.append(
-      NotificationCenter.default.addObserver(
-        forName: NSScrollView.didLiveScrollNotification,
-        object: scrollView,
-        queue: .main
-      ) { [weak self] _ in
-        MainActor.assumeIsolated {
-          self?.handleLiveScroll()
-        }
-      })
-
-    observers.append(
-      NotificationCenter.default.addObserver(
-        forName: NSScroller.preferredScrollerStyleDidChangeNotification,
-        object: nil,
-        queue: .main
-      ) { [weak self] _ in
-        MainActor.assumeIsolated {
-          self?.handleScrollerStyleChange()
-        }
-      })
-
-    observers.append(
-      NotificationCenter.default.addObserver(
-        forName: .ghosttyRuntimeConfigDidChange,
-        object: nil,
-        queue: .main
-      ) { [weak self] _ in
-        MainActor.assumeIsolated {
-          self?.refreshAppearance()
-        }
-      })
-  }
-
-  required init?(coder: NSCoder) {
-    fatalError("init(coder:) is not supported")
-  }
-
-  isolated deinit {
-    for observer in observers {
-      NotificationCenter.default.removeObserver(observer)
-    }
-  }
-
-  override func layout() {
-    super.layout()
-    ensureSurfaceAttached()
-    let effectiveSize = pinnedSize ?? bounds.size
-    scrollView.frame = CGRect(origin: .zero, size: effectiveSize)
-    surfaceView.frame.size = effectiveSize
-    documentView.frame.size.width = effectiveSize.width
-    synchronizeScrollView()
-    synchronizeSurfaceView()
-    surfaceView.updateSurfaceSize()
-  }
-
-  override func viewDidMoveToWindow() {
-    super.viewDidMoveToWindow()
-    ensureSurfaceAttached()
-  }
-
-  func updateSurfaceSize() {
-    surfaceView.updateSurfaceSize()
-    needsLayout = true
-  }
-
-  var isSurfaceAttachedToDocumentView: Bool {
-    surfaceView.superview === documentView
-  }
-
-  func ensureSurfaceAttached(requiresLiveHost: Bool = true) {
-    guard hostKind == .terminal else { return }
-    if requiresLiveHost {
-      guard superview != nil || window != nil else { return }
-    }
-    guard !isSurfaceAttachedToDocumentView else { return }
-    // Only adopt an orphaned surface; never steal it from a live host such as Canvas.
-    guard surfaceView.superview == nil else { return }
-    surfaceLogger.info(
-      "[CanvasExit] hostReattach wrapper=\(debugID) host=\(hostKind.rawValue) "
-        + "surface=\(surfaceView.debugIdentifierForLogging) "
-        + "currentSuperview=\(String(describing: surfaceView.superview)) "
-        + "wrapperWindow=\(window != nil)"
-    )
-    documentView.addSubview(surfaceView)
-    surfaceView.scrollWrapper = self
-    surfaceLogger.info(
-      "[CanvasExit] hostReattachComplete wrapper=\(debugID) host=\(hostKind.rawValue) "
-        + "surface=\(surfaceView.debugIdentifierForLogging) "
-        + "superview=\(surfaceView.superview != nil) "
-        + "window=\(surfaceView.window != nil) "
-        + "bounds=\(Int(surfaceView.bounds.width))x\(Int(surfaceView.bounds.height))"
-    )
-  }
-
-  func updateScrollbar(total: UInt64, offset: UInt64, length: UInt64) {
-    scrollbar = ScrollbarState(total: total, offset: offset, length: length)
-    synchronizeScrollView()
-  }
-
-  func refreshAppearance() {
-    scrollView.hasVerticalScroller = surfaceView.shouldShowScrollbar()
-    scrollView.appearance = NSAppearance(named: surfaceView.scrollbarAppearanceName())
-    scrollView.scrollerStyle = .overlay
-    updateTrackingAreas()
-  }
-
-  private func handleScrollChange() {
-    synchronizeSurfaceView()
-    guard !isProgrammaticScrollChange else {
-      return
-    }
-    updateScrollBackState()
-  }
-
-  private func handleScrollerStyleChange() {
-    refreshAppearance()
-    surfaceView.updateSurfaceSize()
-  }
-
-  private func synchronizeSurfaceView() {
-    let visibleRect = scrollView.contentView.documentVisibleRect
-    surfaceView.frame.origin = visibleRect.origin
-  }
-
-  private func synchronizeScrollView() {
-    documentView.frame.size.height = documentHeight()
-    if !isLiveScrolling && !isUserScrolledBack {
-      let cellHeight = surfaceView.currentCellSize().height
-      if cellHeight > 0, let scrollbar {
-        let targetY =
-          CGFloat(scrollbar.total - scrollbar.offset - scrollbar.length) * cellHeight
-        isProgrammaticScrollChange = true
-        defer { isProgrammaticScrollChange = false }
-        scrollView.contentView.scroll(to: CGPoint(x: 0, y: targetY))
-        lastSentRow = Int(scrollbar.offset)
-      }
-    }
-    scrollView.reflectScrolledClipView(scrollView.contentView)
-  }
-
-  /// Tracks whether the user intentionally moved away from the live bottom of
-  /// the terminal. While this is true we keep the viewport fixed so incoming
-  /// output cannot yank scrollback out from under the user.
-  private func updateScrollBackState() {
-    let cellHeight = surfaceView.currentCellSize().height
-    guard cellHeight > 0 else {
-      isUserScrolledBack = false
-      return
-    }
-
-    let visibleRect = scrollView.contentView.documentVisibleRect
-    let distanceFromBottom = max(0, documentView.frame.height - visibleRect.maxY)
-    isUserScrolledBack = distanceFromBottom > cellHeight / 2
-  }
-
-  private func handleLiveScroll() {
-    let cellHeight = surfaceView.currentCellSize().height
-    guard cellHeight > 0 else { return }
-    let visibleRect = scrollView.contentView.documentVisibleRect
-    let documentHeight = documentView.frame.height
-    let scrollOffset = documentHeight - visibleRect.origin.y - visibleRect.height
-    let row = Int(scrollOffset / cellHeight)
-    guard row != lastSentRow else { return }
-    lastSentRow = row
-    surfaceView.performBindingAction("scroll_to_row:\(row)")
-  }
-
-  private func documentHeight() -> CGFloat {
-    let contentHeight = scrollView.contentSize.height
-    let cellHeight = surfaceView.currentCellSize().height
-    if cellHeight > 0, let scrollbar {
-      let documentGridHeight = CGFloat(scrollbar.total) * cellHeight
-      let padding = contentHeight - (CGFloat(scrollbar.length) * cellHeight)
-      return documentGridHeight + padding
-    }
-    return contentHeight
-  }
-
-  override func mouseMoved(with event: NSEvent) {
-    guard NSScroller.preferredScrollerStyle == .legacy else { return }
-    scrollView.flashScrollers()
-  }
-
-  override func updateTrackingAreas() {
-    for trackingArea in trackingAreas {
-      removeTrackingArea(trackingArea)
-    }
-    super.updateTrackingAreas()
-    guard let scroller = scrollView.verticalScroller else { return }
-    addTrackingArea(
-      NSTrackingArea(
-        rect: convert(scroller.bounds, from: scroller),
-        options: [
-          .mouseMoved,
-          .activeInKeyWindow,
-        ],
-        owner: self,
-        userInfo: nil
-      ))
   }
 }
